Update
I updated the active class in JS, now when you select a numerical rating and hit submit, it will progress to the thank you screen and will display the numerical rating. will have to put safeguards in place to force a rating to be selected before submitting. Will have to research that some more.