Update to all files
Added better visuals in the style.css file to look closer to the figma file. I also had to tweak some of the colors to make them look better. I also set some of the margins and padding as well as typing styles. In the JS file, I added an if/else to the submit button to ensure that it doesn't proceed without a rating.